[js/web] upgrade version of flatbuffers
