+++ js/web/lib/onnxjs/graph.ts
@@ -2,12 +2,10 @@
 // Licensed under the MIT License.
 
 import { Attribute } from './attribute';
-import { onnxruntime } from './ort-schema/flatbuffers/ort-generated';
+import * as ortFbs from './ort-schema/flatbuffers/ort-generated';
 import { onnx } from './ort-schema/protobuf/onnx';
 import { Tensor } from './tensor';
 import { LongUtil, MAX_CLIP, MIN_CLIP, ProtoUtil } from './util';
-
-import ortFbs = onnxruntime.experimental.fbs;
 
 export declare namespace Graph {
   export interface Shape {
