+++ js/web/lib/onnxjs/model.ts
@@ -1,15 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { flatbuffers } from 'flatbuffers';
+import * as flatbuffers from 'flatbuffers';
 
 import { Graph } from './graph';
 import { OpSet } from './opset';
-import { onnxruntime } from './ort-schema/flatbuffers/ort-generated';
+import * as ortFbs from './ort-schema/flatbuffers/ort-generated';
 import { onnx } from './ort-schema/protobuf/onnx';
 import { LongUtil } from './util';
-
-import ortFbs = onnxruntime.experimental.fbs;
 
 export class Model {
   // empty model
