+++ js/web/lib/onnxjs/ort-schema/flatbuffers/onnxruntime/fbs/arg-type.ts
@@ -1,0 +1,8 @@
+// automatically generated by the FlatBuffers compiler, do not modify
+
+/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-explicit-any, @typescript-eslint/no-non-null-assertion */
+
+export enum ArgType {
+  INPUT = 0,
+  OUTPUT = 1,
+}
