+++ js/web/lib/onnxjs/ort-schema/flatbuffers/onnxruntime/fbs/node-type.ts
@@ -1,0 +1,8 @@
+// automatically generated by the FlatBuffers compiler, do not modify
+
+/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-explicit-any, @typescript-eslint/no-non-null-assertion */
+
+export enum NodeType {
+  Primitive = 0,
+  Fused = 1,
+}
