+++ js/web/lib/onnxjs/tensor.ts
@@ -4,11 +4,9 @@
 import { Guid } from 'guid-typescript';
 import Long from 'long';
 
-import { onnxruntime } from './ort-schema/flatbuffers/ort-generated';
+import * as ortFbs from './ort-schema/flatbuffers/ort-generated';
 import { onnx } from './ort-schema/protobuf/onnx';
 import { decodeUtf8String, ProtoUtil, ShapeUtil } from './util';
-
-import ortFbs = onnxruntime.experimental.fbs;
 
 export declare namespace Tensor {
   export interface DataTypeMap {
