+++ js/web/package.json
@@ -10,7 +10,7 @@
   "version": "1.21.0",
   "jsdelivr": "dist/ort.min.js",
   "dependencies": {
-    "flatbuffers": "^1.12.0",
+    "flatbuffers": "^25.1.24",
     "guid-typescript": "^1.0.9",
     "long": "^5.2.3",
     "onnxruntime-common": "file:../common",
@@ -38,7 +38,6 @@
     "@petamoriken/float16": "^3.8.7",
     "@types/chai": "^4.3.4",
     "@types/emscripten": "^1.39.6",
-    "@types/flatbuffers": "^1.10.0",
     "@types/karma": "^6.1.0",
     "@types/minimatch": "^5.1.2",
     "@types/minimist": "^1.2.2",
